feat: add suspense and lazy loading for Dashboard
