+++ src/App.js
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 
 import Header from "components/Header";
-import { Dashboard } from "containers/Dashboard";
+import Dashboard from "containers/Dashboard";
 
 const Container = styled.div`
   width: 100%;
